bug fix for postgres and added tests to check revoke old sessions function

+++ src/Helpers/RefreshToken.php
@@ -84,6 +84,6 @@ class RefreshToken {
     }
 
     public static function getValidity() {
-        return Config::get('supertokens.tokens.refreshToken.validity') * 60 * 60;
+        return (int)(Config::get('supertokens.tokens.refreshToken.validity') * 60 * 60);
     }
 }
